Add files via upload
In this project we use pandas for data cleaning. Use matplotlib and seaborn to create plot to better visualization of data based on the different country. which is the most popular video game and over different age group and over the year does popularity  increase or decrease of video games.